Do not run db cleanup in CI

+++ run_test.sh
@@ -5,12 +5,14 @@ err=0
 trap 'err=1' ERR
 
 # Clean
-if psql -lqt | cut -d \| -f 1 | grep -qw hello_world ; then
-    read -p "Database 'hello_world' required for running the tests already exist. Do you want to delete it (y)?" yn
-    if echo "$yn" | grep -iq "^n" ;then
-        exit
-    else
-        dropdb hello_world
+if [ ! $CI ]; then
+    if psql -lqt | cut -d \| -f 1 | grep -qw hello_world ; then
+        read -p "Database 'hello_world' required for running the tests already exist. Do you want to delete it (y)?" yn
+        if echo "$yn" | grep -iq "^n" ;then
+            exit
+        else
+            dropdb hello_world
+        fi
     fi
 fi
